- Integer to Long - Added generator readme - Changed readme of demosql and demopsql - Changed port of data-generator

+++ demomysql/src/main/java/xyz/hemoreg/demomysql/usermanagment/controller/ClientController.java
@@ -24,7 +24,7 @@ public class ClientController {
     }
 
     @GetMapping(path= "{id}")
-    public ClientDTO getClientById(@PathVariable Integer id) {
+    public ClientDTO getClientById(@PathVariable Long id) {
         return clientService.getClientById(id);
     }
 }
